requirements.txt
praw==7.8.1
requests==2.32.3
python-dotenv==1.0.1
